requirements.txt
six
django
django-filter

